+++ src/handler/interfaces/IErrorHandlingStrategy.ts
@@ -1,0 +1,5 @@
+import { Response } from "express";
+export interface IErrorHandlingStrategy {
+  canHandle(err: Error): boolean;
+  handle(err: Error, res: Response): void;
+}