+++ src/index.ts
@@ -10,3 +10,10 @@ export * from "./errors/input-validation/zodErrorSerializer.js";
 //serializers
 export * from "./serializers/interfaces/IHttpErrorSerializer.js";
 export * from "./serializers/HttpErrorSerializer.js";
+
+//handler
+export * from "./handler/factory.js";
+export * from "./handler/ErrorHandlerContext.js";
+export * from "./handler/HttpErrorHandlerStrategy.js";
+export * from "./handler/InputValidationErrorHandlerStrategy.js";
+export * from "./handler/FallBackErrorHandlerStrategy.js";